2007-09-10  Dan Williams  <dcbw@redhat.com>
	* src/NetworkManagerAP.c
	  src/NetworkManagerAP.h
	  introspection/nm-access-point.xml
		- Change 'freq' property to a guint32 instead of a double since we
			weren't using the floating point bits anyway



git-svn-id: http://svn-archive.gnome.org/svn/NetworkManager/trunk@2780 4912f4e0-d625-0410-9fb7-b9a5a253dbdc

+++ src/NetworkManagerAP.h
@@ -89,8 +89,8 @@ void				nm_ap_set_mode			(NMAccessPoint *ap, const int mode);
 gint8			nm_ap_get_strength		(NMAccessPoint *ap);
 void				nm_ap_set_strength		(NMAccessPoint *ap, gint8 strength);
 
-double			nm_ap_get_freq			(NMAccessPoint *ap);
-void				nm_ap_set_freq			(NMAccessPoint *ap, double freq);
+guint32			nm_ap_get_freq			(NMAccessPoint *ap);
+void				nm_ap_set_freq			(NMAccessPoint *ap, guint32 freq);
 
 guint16			nm_ap_get_rate			(NMAccessPoint *ap);
 void				nm_ap_set_rate			(NMAccessPoint *ap, guint16 rate);
